COMMON: Concurrency: Introduce a RCU Like primitive for Atomic Uniq Ptrs
We introduce a simple Kernel RCU like primitive to track readers in their
critical sections. The idea is for rarely changing data, that readers
do not wait upon writers. The readers track when they enter and leave critical
sections and the writer side does not delete the old data until we're sure that
no readers access this.

The idea here is that whenever a publisher updates the data we increment the
epoch of the data, the epoch is incremented post an atomic ptr reset, and this
ensures that no threads will access data newer than the published epoch, and
this idea can be used to reclaim the older epochs when the count of threads
accessing epochs is zero. We introduce 2 types of epoch counter

- A simple epoch counter where every thread gets the current epoch of the data
and increment their counts, the difference between this approach vs RW Locks is
that we still keep alive the old version of data while new data is published, so
readers are never blocked even if a writer comes along. However all threads
almost read the same memory, so this has high cache contention.

- A thread local epoch counter: here a thread local TID is generated similar to
folly and CMU's deferred RCU approach. This is done statically, so the first
time the thread tries to do some locking a thread local TID is generated which
is indexed onto a static array. The array elements are cache aligned, so for
readers we should have the ideal case of no sharing as far as the locking
primitive itself goes, only the writer has to pay the cost of walking through
this array to determine the epochs

Tests are added to demonstrate the various functionalities, and microbenchmarks
follow to demonstrate the different gains

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>