MGM: placement: add print and duplicate detection for PlacementResult
Add a simple contains() member function that checks if the result already
contains a specific item, while this is traditionally done via a set for eg. in
this case the array is so small that iteration should win. RandomStrategy uses
this as it is possible that a particular item might already be picked

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>