eoscp: Added checksum comparison between the source and the destination
Only works if the destination is an XRootD server.

If one provides the -C option to eoscp, a checksum comparison will be
done between the source and the destination. If the checksums mismatch,
an error will be displayed on stderr.
If one provides -E in addition to the -C option, if the checksums
mismatch, the file will be deleted on the destination side.

Also includes a fix containing checksum-type label alignment