MGM: GroupBalancer: introduce a freespace engine
This introduces a new groupbalancer engine called freespace engine that tries to
make the absolute freespace of every group equal. There are tuning knobs to
control how much of a spread is around using the classic min/max threshold in
order to avoid jitters once we are nearly equal. The default value is 2%, which
means for eg. if freespace per group is ideally 50TB we allow for values between
49 and 51. Also introducing a new feature whereby we allow for blocklisting
groups that participate in the groupbalancer, this only applies to the freespace
engine, where the config option space.groupbalancer.blocklist option can take a
commalist of groups that do not participate in the group balancing activity

Fixes: EOS-5803
Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>
(cherry picked from commit 4462878673c63535e384a2809119844d7a3f8025)

Conflicts:
	mgm/proc/admin/SpaceCmd.cc
dropping new scheduler options only present in master