NS: Attr - WIP attempt to call find multiple time to prevent race for Attr recursive operations
It's unfortunately very slow compared to the performing of the big find
at the beginning of the command. For 50k directories, it takes around 20
seconds to list the extended attributes recursively with the initial
implementation. With the new implementation, it takes around 30 seconds
--> all the find take 16 seconds in total compared to around 10 seconds
for the run of the unique find.