FST: FSCK: Update stripe_err check to detect overreplication in rain files
When checking for stripe_err, now verify all the different locations,
taking into account duplicate stripes. If the file is overreplicated,
mark duplicated stripes as invalid.

The RainMetaLayout::OpenPio had to be extended to make possible openning
a stripe with a replica index greater than the number of stripes in the
layout.

Overreplicated stripes for rain layouts are now dropped in the fsck repair job.

Add a function to copy files between containers for tests, preserving their xattrs.