COMMON: Logging: check for active_head null inside the lock
Accessing `LogBuffer::active_head` outside of the lock is unsafe, another thread
could be updating this value while we're trying to read, so check for nullness
before unlock. The modification doesn't protect 2 threads trying to flush at the
same time, this could happen before this change as well, however this
is (likely) rare. In case we see issues with flush, we need to introduce another
mutex for protecting stderr

Fixes: EOS-5944
Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>