FUSEX: mdstackfree: drop orphaned inodes only if we manage to get a lock
This changes the behaviour of mdstackfree routine where we try to lock the
metadata object to get its parent id as this needs to be consistent. However in
order to avoid deadlocks we do a try_lock and do this only if we manage to
atomically grab a lock. There are other places that call locking in an inverted lock
order, for eg: unlink which calls retrieveTS where the local mdlock is held first before
a mdmap lock. In order to be consistent also the mdmap cannot be unlocked as this
can lead to situations where the map has changed while we unlock so the next routines are
no longer valid. So we hold the map lock, atomically try to lock the shared md object,
and proceed with removal if we manage or continue this in the next cycle.

The parent id is made an optional so that we proceed
only when the value is set or try this in the next loop

Fixes: EOS-5965
Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>