NAMESPACE: Prefetcher: drop double calls of future objects
Drop unused future objects, the code is called when we create the URI objects
anyway. Also the folly::future version we use in the current version is unclear
on the safety of moving futures as this version explicitly mentions futures
splitting in two lines may not be concurrency safe.

Since the actual call site that uses the future is the one calling the
continuation use this call only. Look at EOS-5931 for some details.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>