MGM: placement: ClusterDataTypes: provide helpers for translating group iter ids
We internally use a negative number to represent a group so that we can walk
down the tree and understand conditional endpoint once come across +ve IDs.
However this output is confusing as an ID, so write the expected group ID also
with the bucket output. Also improve the error/info messages logged with these IDs

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>