NS: Added the try-lock type of Locker class for Container/FileMD - NSObjectTryLocker
Also adapted the BulkNSObjectLocker class that will loop until it could
lock all the objects that it needs to lock. If one object cannot be
locked, the locks already taken will be released and the locking of all
the objects will be retried again.
This fix prevents potential deadlocks that can happen when directories
with the same parents are renamed at the same time of the listing of the
extended attributes of the parent.