CMAKE: Drop requirement of librocksdb_tools.so
When linking with the static libraries, some symbols are missing
in librocksdb.a, so librocksdb_tools.a needs to be linked as well.
When using shared libraries, all symbols can be found in the main
librocksdb.so, so it's not necessary to link the tools library.
Moreover, the tools library is not usualy installed as part of most
rocksdb packages in distributions.