MGM: Send deletion event for FileSystem removal
Make sure the FileSystem Shared Hash object is deleted when the file
system is removed and the deletion event gets sent to the FST
otherwise we end up with stale data structures on the FST. This
in turn can prevent the registration of a new file system at the
same location later on. Ignore configuration changes at the master
MGM otherwise we end up with race conditions when deleting and
adding the same type of objects e.g. file systems.

Fixes EOS-6101