CONSOLE: eos-iam-mapfile: support oneshot gridmap generation
Since usually one needs to generate gridmap from various IAM endpoints with
different mappings, do this in the script itself, this allows for the script to
be invoked similar to mkgridmap, where various mappings and localmapping can be
specified. The IAM config now supports an account key, which is the account
mapping to be done for that endpoint.

Currently the intermediary files aren't deleted at the end for debug, this will
be done at a future stage.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>