CMAKE: Update CMake minimum requirement and supported versions
Note that, due to CMP0119, we needed to set the language of EosBlake3
sources to the correct language, otherwise CMake compiles it as C and
fails. This change in behavior was introduced in CMake 3.20. In order
to compile it as assembly, we need to add ASM to the list of enabled
languages in the call to project(), as only C and CXX are enabled by
default.

https://cmake.org/cmake/help/v3.20/policy/CMP0119.html