NS: Removed file lock before calling HierarchicalView::getUri()
This prevents deadlock from happening.
If one read/write locks a file before calling eosView->getUri(fmd), the
logic of getUri() will try to lock the parent container of the file in
order to access the file parent's parent ID. It will therefore not
respect the locking order that should be 1. lock(Container) 2.
lock(File)