NS: ContainerAccounting - Refactored the ContainerAccounting logic for new NS locking implementation
The aim was to prevent that QueueForUpdate() goes up the tree. The tree
browsing from bottom to up is now done in a separate thread.
As an example, this allows to call QuarkFileMD::setSize() without
risking to deadlock if the parent container was not locked before
calling setSize().

There are now two threads in the ContainerAccounting logic. One that
will be responsible for browsing the tree to update from the container
to '/' (AsyncQueueForUpdate) and submit it to the second thread that will be responsible of
updating the entire tree size (PropagateUpdate).

Also removed NS mutex variable as it is not used anymore.