COMMON: Added AddEosApp() function to add eos.app opaque to a URL
This function was created in order to add eos.app=xyz to a URL or a
string containing opaque infos.

If eos.app is already present in the pathOrOpaque, the value will be prepended
by protocol/. If the value is equal to protocol, then nothing will be done
If the value is equal to protocol/xyz, then nothing will be done
e.g: pathOrOpaque=/eos/test/fic.txt?eos.app=hello --> pathOrOpaque will be equal to
/eos/test/fic.txt?eos.app=protocol/hello
e.g2: pathOrOpaque= test=test1&eos.app=protocol --> pathOpaque will be left untouched
e.g3: pathOrOpaque=/eos/test/fic.txt?eos.app=http/s3 --> pathOpaque will be left untouched
If eos.app is not already present in the pathOrOpaque, the value of eos.app
will be eos.app=protocol
If eos.app is provided twice in the pathOrOpaque, only the last occurence of eos.app will be
considered. Indeed, usually the opaque will be transformed into a XrdOucEnv and it
is implemented using a hash --> the second occurence will overwrite the first one!

Additionally added unit tests.