COMMON: Mapping: use getgrouplist to simplify fetching secondary groups
A new interface is introduced which allows for fetching groups via
getgrouplist insted of getgrent which are rentrant. A env allows for this to be
configurable at runtime

The defaults have changed to getting grouplist, which means for a given user we
only fetch their groups with this, drastically reducing the calls to a stateful
getgrent and co which might be heavy and cause errors with lots of queries on
LDAP/SSSD backends.

Fixes: EOS-6356
Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>
(cherry picked from commit af6cd9cf7b3fd94da70abc5732b343ed89e2594d)

 Conflicts:
	* common/CMakeLists.txt
  Dropped BehaviourConfig
	* common/Mapping.hh
	* unit_tests/common/MappingTests.cc
  Header include format changes in 5.3