FST: Return EAGAIN instead of EBUSY in case a file is opened for writing via HTTP
This is a hack against the hack in XRootD in the Emsg() function.
We initially wanted to use EBUSY. However, if EBUSY errno is passed in Emsg(), the Emsg() function in XRootD
will return the integer 5 (Hack for proxy support).
This is going to explose when this hack is removed in XRootD 6 hence the
decision to use EAGAIN