MGM: Scheduler/GeoTreeEngine: make existingReplicas const-ref/ptr param
We don't modify the locationsfs/existingReplicas parameter in GeoTreeEngine
or Scheduler::Access. Make this a pointer to const or a const ref
to ensure that we don't do this. All the vector iterator methods
are moved to their const variants to comply. printf functions
are also changed to %u to indicate unsigned to be in line with
fsid_t definitions everywhere.
fsIndex is now using std::distance for ptr arithmetic instead of a wrong cast

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>