MGM: Open: move proxy/fw settings to a seperate function
The proxy/firewall settings are moved to a helper function
`setProxyFwEntrypoint` which handles both firewalls and proxy settings. It
returns an output struct that has these values filled and evaluates to true only
when there is an actual modification which can be used to distinguish
the presence/absence of proxys/gateways.

Also the XrdOucString params are changed to std::string using the
common::replace_all where applicable when proxys are set.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>