COMMON: StringUtils: introduce a XrdOucString equivalent replace_all
This function would replace all src patterns within the string with a
target. When the target replacement is larger than source
the output string is likely going to be a larger string, in this case
we do an optimisation pass to determine the total size we'd grow to
and reserve the string. This allows for string::replace function calls
to never realloc while maintaining the similar memmove/memcpy
patterns that'd be eventually called with std::string::replace functions

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>